requirements.txt
apscheduler~=3.11.0
config~=0.5.1
python-telegram-bot~=21.9
requests~=2.32.3
flask~=3.1.0
python-dotenv~=1.1.0
cachetools>=6.1.0
h11~=0.14.0
pip~=25.3
attrs~=24.2.0
PySocks~=1.7.1
cryptography~=44.0.0
pillow~=11.0.0
pytz~=2024.2
sniffio~=1.3.1
trio~=0.30.0
outcome~=1.3.0.post0
click~=8.1.7
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
certifi~=2024.8.30
tzlocal~=5.2
anyio~=4.12.1
telegram~=0.0.1
six~=1.17.0
gevent~=24.11.1
pymongo~=4.10.1
beautifulsoup4~=4.14.3
distro~=1.9.0
filelock~=3.20.3
huggingface_hub~=1.3.2
selenium~=4.22.0